requirements.txt
# ===== Basic Dependencies =====
streamlit>=1.35.0
fastapi>=0.100.0
uvicorn>=0.20.0
langchain-core>=0.1.30
langchain-community>=0.0.20
langchain-deepseek
langgraph>=0.0.25
pydantic-settings
rich
requests
openai

# ===== RAG Dependencies (for memory retrieval) =====
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
# Note: Use faiss-gpu instead of faiss-cpu if GPU is available

# ===== Phase 1: Cache + Learning Profile =====
# pymysql - optional, only needed when PROFILE_STORE=mysql

# ===== Phase 2: Knowledge Graph =====
networkx>=3.0
pyvis>=0.3.0
plotly>=5.18.0
transformers>=4.36.0
torch>=2.0.0
keybert>=0.8.0
spacy>=3.7.0
# After installation, download Chinese model:
# python -m spacy download zh_core_web_sm

# ===== Desk Pet (PyQt6 Desktop Pet) =====
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
pyaudio>=0.2.13

# ===== macOS Desktop Pet Support (pyobjc) =====
# Only required on macOS for window management
# pyobjc>=9.0
# pyobjc-framework-Quartz>=9.0
# pyobjc-framework-Cocoa>=9.0
